+++ src/main/java/com/yeoljeong/tripmate/notification/domain/model/NotificationHistory.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.notification.domain.entity;
+package com.yeoljeong.tripmate.notification.domain.model;
 
 import com.yeoljeong.tripmate.domain.BaseAuditEntity;
 import com.yeoljeong.tripmate.notification.domain.constants.ChannelType;
